Building the arena works, getting the numa node too, pinning still gives ENOMEM

+++ include/config.hpp
@@ -1,4 +1,4 @@
-#include <cstddef>
+#include <cstdint>
 #define EXT_HEAP_SZ_EXP 25
 
-size_t ext_heap_sz = 1 << 25;
+uint32_t ext_heap_sz = 1 << 25;
